+++ .storybook/preview.ts
@@ -1,8 +1,0 @@
-import type { Preview } from "@storybook/react";
-import "./globals.css";
-
-const preview: Preview = {
-  parameters: { layout: "centered" },
-};
-
-export default preview;